fix: Temporarily disable Flagger until infrastructure is deployed
Deployment was failing because Flagger CRDs are not installed yet:
'no matches for kind "Canary" in version "flagger.app/v1beta1"'

The infrastructure workflow needs to run first to install Flagger.
Disabled Flagger in production values to allow initial deployment.

Steps to re-enable Flagger:
1. Run infrastructure workflow to deploy Flagger CRDs
2. Set flagger.enabled: true in values-production.yaml
3. Redeploy application

This is a temporary fix for first-time production deployment.